Added mouse_agent
mouse_agent senses the action space and uses the mouse for continuous
spaces and the keyboard for discrete spaces. Also supports the ‘q’ key
for quitting. Only support continuous shape=(1,) and shape = (2,)
spaces. The horizontal position in the window maps to the first
dimension of the space, and the vertical mouse position maps to the
second dimension.